requirements.txt
torch==2.0.0
torchvision==0.15.1
timm==0.6.7
numpy
Pillow
opencv-python
plyfile
tensorboardX
einops